Reorganize, cleanup and add copyright
